int to double

+++ RfCommonLibrary/Recipes/DTOs/QueryDTOs/IngredientDTO.cs
@@ -1,12 +1,7 @@
 ﻿using RfCommonLibrary.Recipes.DatabaseModels;
 using RfCommonLibrary.Recipes.Enums;
-using System;
-using System.Collections.Generic;
 using System.ComponentModel.DataAnnotations;
-using System.Linq;
-using System.Text;
 using System.Text.Json.Serialization;
-using System.Threading.Tasks;
 
 namespace RfCommonLibrary.Recipes.DTOs.QueryDTOs
 {
@@ -36,7 +31,7 @@ namespace RfCommonLibrary.Recipes.DTOs.QueryDTOs
         [StringLength(25, MinimumLength = 2, ErrorMessage = "MIN: 2, MAX: 25")]
         public string Name { get; set; } = string.Empty;
         [Required]
-        public int Amount { get; set; } = 0;
+        public double Amount { get; set; } = 0;
         [Required]
         public MeasurementType Unit { get; set; } = MeasurementType.Unknown;
         [Required]
